1. update node to v16; 2. add pcb store; 3. auth/login logic ui adjust; 4. apply api data in login and wirings

+++ src/share/auth/auth.service.ts
@@ -1,7 +1,7 @@
-import { YNCacheKey, cacheService, httpService, StorageType } from "@/share";
-import { APP_URL } from '@/share';
-import { of } from "rxjs";
-import { map, tap } from "rxjs/operators";
+import { YNCacheKey, cacheService, httpService, StorageType, YNAPI_System } from "@/share";
+import { of, throwError } from "rxjs";
+import { catchError, map, switchMap, tap } from "rxjs/operators";
+import { alertService } from "../alert.service";
 
 export interface RequestUserInfo {
   userName: string;
@@ -11,34 +11,66 @@ export interface RequestUserInfo {
 export interface ResponseUserInfo {
   userId: string;
   userName: string;
-  roleId: number;
+  loginUserName: string;
+  email: string;
   remenberPassword: boolean;
   password: string;
+}
+
+interface UserResponseInfo {
+  custom: any
+  message: string;
+  success: boolean;
 }
 // remenber true, userName and password save in localStorage
 // remenber false, userName and password save in sessionStorage
 
 class AuthService {
   login(user: RequestUserInfo) {
-    return httpService.post<ResponseUserInfo>(APP_URL.Login, user).pipe(
-      map(() => {
-        const responseUser: ResponseUserInfo = {
-          userId: '_user_id_1_',
-          userName: user.userName,
-          remenberPassword: user.remenberPassword,
-          roleId: 789789,
-          password: user.remenberPassword ? user.password : '',
-        };
-        // need remove, may user data store in diff storage in prev and next; 
-        cacheService.remove(YNCacheKey.User);
-        cacheService.set(YNCacheKey.User, responseUser, user.remenberPassword ? StorageType.Persistent : StorageType.Session);
-        cacheService.set(YNCacheKey.AccessToken,  '__test_token__',  StorageType.Session);
-        return responseUser;
-      })
+    return httpService.post<{data: UserResponseInfo}>(YNAPI_System.Login, {loginUserName: user.userName, pwd: user.password }).pipe(
+      catchError(error => {
+        return of({
+          data: {
+            success: false,
+            message: (error?.message?.toString()|| '') as string ,
+            custom: {}
+          },
+        });
+      }),
+      map(response => {
+        let responseUser: ResponseUserInfo | null = null;
+        if (response?.data?.success) {
+          const token = response.data.custom.Token;
+          responseUser = {
+            userId: response.data.custom.UserID,
+            userName: response.data.custom.UserName,
+            loginUserName: response.data.custom.LoginName,
+            email: response.data.custom.Email,
+            remenberPassword: user.remenberPassword,
+            password: user.remenberPassword ? user.password : '',
+          };
+          
+          // need remove, may user data store in diff storage in prev and next; 
+          cacheService.remove(YNCacheKey.User);
+          cacheService.set(YNCacheKey.User, responseUser, user.remenberPassword ? StorageType.Persistent : StorageType.Session);
+          cacheService.set(YNCacheKey.AccessToken,  token,  StorageType.Session);
+        }
+        return {
+          success: response.data.success,
+          message: response.data.message || '',
+          user: responseUser
+        };        
+      }),
+
     );
   }
   logout() {
-    return of(cacheService.remove(YNCacheKey.AccessToken));
+    return httpService.post<ResponseUserInfo>(YNAPI_System.LogOut).pipe(
+      tap(() => {
+        cacheService.remove(YNCacheKey.AccessToken); 
+      })
+    );
+    // return of(cacheService.remove(YNCacheKey.AccessToken));
   }
 }
